Move controller metrics boilerplate to a separate file
