+++ pkg/controller/baremetalhost/baremetalhost_controller.go
@@ -20,8 +20,6 @@ import (
 
 	"github.com/go-logr/logr"
 
-	"github.com/prometheus/client_golang/prometheus"
-
 	corev1 "k8s.io/api/core/v1"
 
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -35,43 +33,21 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
-	"sigs.k8s.io/controller-runtime/pkg/metrics"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
 )
 
 const (
 	hostErrorRetryDelay = time.Second * 10
-
-	labelHostNamespace = "namespace"
-	labelHostName      = "host"
-	labelPowerOnOff    = "on_off"
 )
 
 var runInTestMode bool
 var runInDemoMode bool
-var reconcileCounters = prometheus.NewCounterVec(prometheus.CounterOpts{
-	Name: "metal3_reconcile_total",
-	Help: "The number of times hosts have been reconciled",
-}, []string{labelHostNamespace, labelHostName})
-var reconcileErrorCounter = prometheus.NewCounter(prometheus.CounterOpts{
-	Name: "metal3_reconcile_error_total",
-	Help: "The number of times the operator has failed to reconcile a host",
-})
-var powerChangeAttempts = prometheus.NewCounterVec(prometheus.CounterOpts{
-	Name: "metal3_operation_power_change_total",
-	Help: "Number of times a host has been powered on or off",
-}, []string{labelHostNamespace, labelHostName, labelPowerOnOff})
 
 func init() {
 	flag.BoolVar(&runInTestMode, "test-mode", false, "disable ironic communication")
 	flag.BoolVar(&runInDemoMode, "demo-mode", false,
 		"use the demo provisioner to set host states")
-
-	metrics.Registry.MustRegister(
-		reconcileCounters,
-		reconcileErrorCounter,
-		powerChangeAttempts)
 }
 
 var log = logf.Log.WithName("baremetalhost")
@@ -166,10 +142,7 @@ func (info *reconcileInfo) publishEvent(reason, message string) {
 // is true, otherwise upon completion it will remove the work from the
 // queue.
 func (r *ReconcileBareMetalHost) Reconcile(request reconcile.Request) (result reconcile.Result, err error) {
-	reconcileCounters.With(prometheus.Labels{
-		labelHostNamespace: request.Namespace,
-		labelHostName:      request.Name,
-	}).Inc()
+	reconcileCounters.With(hostMetricLabels(request)).Inc()
 	defer func() {
 		if err != nil {
 			reconcileErrorCounter.Inc()
@@ -632,10 +605,7 @@ func (r *ReconcileBareMetalHost) manageHostPower(prov provisioner.Provisioner, i
 
 	if provResult.Dirty {
 		info.postSaveCallbacks = append(info.postSaveCallbacks, func() {
-			metricLabels := prometheus.Labels{
-				labelHostNamespace: info.host.Namespace,
-				labelHostName:      info.host.Name,
-			}
+			metricLabels := hostMetricLabels(info.request)
 			if info.host.Spec.Online {
 				metricLabels[labelPowerOnOff] = "on"
 			} else {
