+++ pkg/controller/baremetalhost/host_state_machine.go
@@ -6,40 +6,8 @@ import (
 	metal3v1alpha1 "github.com/metal3-io/baremetal-operator/pkg/apis/metal3/v1alpha1"
 	"github.com/metal3-io/baremetal-operator/pkg/provisioner"
 
-	"github.com/prometheus/client_golang/prometheus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"sigs.k8s.io/controller-runtime/pkg/metrics"
 )
-
-var slowOperationBuckets = []float64{30, 90, 180, 360, 720, 1440}
-var stateTime = map[metal3v1alpha1.ProvisioningState]*prometheus.HistogramVec{
-	metal3v1alpha1.StateRegistering: prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Name: "metal3_operation_register_duration_seconds",
-		Help: "Length of time per registration per host",
-	}, []string{labelHostNamespace, labelHostName}),
-	metal3v1alpha1.StateInspecting: prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Name:    "metal3_operation_inspect_duration_seconds",
-		Help:    "Length of time per hardware inspection per host",
-		Buckets: slowOperationBuckets,
-	}, []string{labelHostNamespace, labelHostName}),
-	metal3v1alpha1.StateProvisioning: prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Name:    "metal3_operation_provision_duration_seconds",
-		Help:    "Length of time per hardware provision operation per host",
-		Buckets: slowOperationBuckets,
-	}, []string{labelHostNamespace, labelHostName}),
-	metal3v1alpha1.StateDeprovisioning: prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Name:    "metal3_operation_deprovision_duration_seconds",
-		Help:    "Length of time per hardware deprovision operation per host",
-		Buckets: slowOperationBuckets,
-	}, []string{labelHostNamespace, labelHostName}),
-}
-
-func init() {
-	for _, collector := range stateTime {
-		metrics.Registry.MustRegister(collector)
-	}
-}
 
 // hostStateMachine is a finite state machine that manages transitions between
 // the states of a BareMetalHost.
@@ -98,10 +66,8 @@ func recordStateEnd(info *reconcileInfo, host *metal3v1alpha1.BareMetalHost, sta
 		if !prevMetric.Start.IsZero() {
 			prevMetric.End = time
 			info.postSaveCallbacks = append(info.postSaveCallbacks, func() {
-				stateTime[state].With(prometheus.Labels{
-					labelHostNamespace: host.Namespace,
-					labelHostName:      host.Name,
-				}).Observe(prevMetric.Duration().Seconds())
+				observer := stateTime[state].With(hostMetricLabels(info.request))
+				observer.Observe(prevMetric.Duration().Seconds())
 			})
 		}
 	}
